data-container: throw a warning when two or more data-containers are the recipients of a loaded page
When an href='#/..' is clicked, the target element to load the content should be unique. When there are more container candidates for that href, the content is loaded into the wrong container by executing Murphy's law. A console warning is useful to detect unwanted containers.

Even thought this issue is usually a minor one, it could lead to start debugging the library waiting for errors, and no errors are fired, because the content is loaded into the container perfectly... well... into the wrong container...