template: change CSS display mode for unprocessed templates (from visibility to display)
The previous version had visibility hidden or visible depending on the data-status of the template.
This mode was good for trying to fit the space before processing the template but it has a visual glitch when anidated templates are processed before the main template. In order to solve this, CSS display property is set, so the view will be displayed only whe the parent template has been processed (templates of same level 1 will still rendering independently, because there is no issue with them).