router:data-load: handy property to load pages by using <tag.. data-load='url'>
Instead of using javascript to load pages by using ir.router, this is an "easy to load" way.
Please note that this attribute should be used in snippet pages better than in the main page because data-load is only loaded along with the requested page. This functionality could be added in future implementations