template: allow implicit 'this' on {{method()}}
Methods from the controller can be written like this <button onclick="{{update()}}">

Remember that these methods return a promise, so you can chain them.

In this case update() is valid without knowing 'this'

But if method is , for instance, remove(), we need to pass the row to be removed.

Instead of calling {{remove(element)}} we use {{remove()}} and the row is later guessed from the this object (the element that fired the click)

The idea is that all methods will have a this object with info about the context, so the syntax is easier.

You can always call directy by using

- ir.controller('templateName').remove(this.parent)

but

- {{remove()}}

is short and it has two implicit variables (the template name and the element)