controller: remove() method, add() is not implemented yet
Instead of CRUD methods (useful with server) these new methods can modify the model and the view without sending data to the server. The server update would be performed when the user is sure of the changes and press a button called "save". This button will call the controller 'update' method.

Remove is pretty useful when working with  Array type models.

Add will be implemeted in next versions