template:paintToTemplate can be called at any time, and the new nodes will be inited
Before, the {{}}elements were processed when init the template, and calling again paintToTemplate just updated the processed elements. Any new element had to be processed manually.

Now, isTemplateInited has been removed from processing elements, so if a new element is inserted, the parser will init it, and if it is already parsed,it will simply update the content.