template:external subscriptions: create atomic subscriptors instead of all the elements in the TemplateName
Before: when a [B element]{{modelA:var1}} was found, the subscription created was controllerB subscribes to modelA changes. Not only the element, but all the elements in the template B. This is easy to implement but is not a good performance solution.

Now: when a {{modelA:var1}} was found, the subscription created is THAT_ELEMENT subscribes to modelA changes. Harder to implement but runtime performance is better due to less DOM operations.

Note: the element must live  inside any template (you can create a template with no model just by adding data-model attribute). If no template containing that element, the default template is used.