template:external subscriptions: autogenerate iridium-default in body for first load, prepare for non-datamodel parsing, and model.obj is a getter,setter to keep track of the creation of data
An external {{a:b}} expression should be always inside a data-model. To avoid creating a model for the main container (and usually forget to put it), a data-model="iridium-default" is added to the <body> tag

Regarding the non required data-model for external expressions. This is not trivial,so just some changes to prepare it for future implementation,

model.obj has been improved to getter and setter, to notify when the object is created for the first time, or when the full object changes.