template: solve problem when <select> has not a null option
The problem was that if a select retrieves values from a data-model, the model could retrieve only data without a default-empty option.

The solution is to add a ir.controller().fn=function(arr){arr.push('-')} and set data-options='fn' in the template

Another solution is to create the template inside the select (an option for the default value and then the template. But the new solution is better.