template:data-options: filtering or sorting retrieves the indexes of modified items (instead of changing the original object)
We lose the ability to add elements from data-options, but the behaviour is better since the model is never modified, and the function applies to the existing items, no more, no less. Sorting and filtering will be really easy to implement by using very simple functions

ir.controller('aaa').sort(arr=>arr.sort) and the new html list is rendered while the model still contains the same original elements (inmutability)

The only problem is that data-options is most useful for arrays and not for objects. But that will change in the future.