template: allow boolean attributes (checked, autofocus...)
Boolean attributes are special DOM attributes. The attribute is executed if it is present, and only removing that property prevents execution (instead of changing the value)

The library can use brackets notation to set true or false. If true, the attribute is added, and if false removed.

Example:

`<input checked='{{isChecked}}'...>`

 and the output is

`<input...>`

or

`<input checked ...>`