general: getObjectProperty() sends console.warn() when value is undefined (not found in the model)
Sometimes, a typo can be inserted in the code and debugging what is the error can take some precious time. This function will output the key when the assigned value is either undefined (the programmer should better assign it no null) or when the key is not found.

Notifying the potential error is much better than failing silently (and swallow the error for a long time until the customer sends an incidence in a  very rude language!).