controller.extend: add custom functionality to controller in a nicer fashion than ir.controller.myFunction=function(){}
If lots of functions, the old method is tedious

Instead, if all functions are declared in the same place, the javascript seems more elegant.

ir.configure(,,,functions) can handle that but sometimes it is better to declare url and options in the data-provider and data-options, so configure is not needed. This is the case for extend() method. Besides, this method returns the controller so you can do ir.controller.extend(...).configure(...)

So to summarize:

- if you prefer to use html data-provider and data-options, use extend() to declare functions or controller.func=...
- if you prefer to use javascript-only use extend().configure() or configure(,,,)