general: MAJOR, replace _load()[jquery] by _load()[vanilla js]
(Tests passed OK)

This function is a plain, tiny, promise and fetch based and using only plain vanilla javascript.

This function will allow to preprocess loaded pages to change relative urls or other things, in an easier way than using regex because we have the DOM elements created, instead of just text.

The old function has been renamed but not deleted because we need to perform more tests (unique scenarios) and the old one was very stable on that.

A note of caution, these two functions are not equivalent because the jquery one uses internaly lazy loading (ajax) with the script, while we invoke them in the same place they are declared.

This is a big step to disconnect the library from the jquery dependency. `load()` is, by far, the most difficult function to emulate. After implemented, it seems that it is a very tiny and straightforward implementation, but it is very easy to say that after done.

Even if this is the breaking point to start to disuse jquery into the library, this is only to reduce the total size of the library from 5KB+31KB to 5KB, just that. Jquery is an incredible library, and I have to thank this library to simplify the js development all around the world with a set of easy but powerful functions.

Probably, lots of people will keep using JQuery in their development, even when 'gurus' are fostering not to use it. You can do the same js with less code, less code means easier to read and less errors. Ah and jquery behaves the same on all browsers, even the old ones!

So instead of saying:

> To use Iridium you don't need JQuery

we can say:

> To use Iridium, JQuery is *optional*.

Thanks JQuery!