controller:  _fireChanges() returns only the controller
Simplify thinking about then() parameters. The result is ALWAYS the controller. Period.

Since the functions calling indirectly that method contains the variables already, they can be used since the thenable function is a closure, so no need to pass new parameters to the thenable function.

The controller could also be removed since it is the 'this' object in the thenable function but it is handy to return it because the controller name is usually aliased.

Example:

```javascript
const testApp = ir.controller('test');

testApp.whatever = function() {
  const message = 'hello'
  this.update().then(controller => console.log(message))
  //this.update().then(testApp => console.log(message))
  //this.update().then(() => console.log(message))
}
```